refactor: backend; file service and create property endpoint

+++ Back End/src/main/java/com/api/globalState/dtos/response/UserResponseDto.java
@@ -5,7 +5,7 @@ import lombok.Data;
 
 @Data
 public class UserResponseDto {
-    private Long idUser;
+    private Integer idUser;
 
     private String name;
