+++ Back End/src/main/java/com/api/globalState/entities/auth/UserEntity.java
@@ -17,7 +17,7 @@ public class UserEntity {
 
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Id
-    private Long idUser;
+    private Integer idUser;
 
     private String name;
